requirements.txt
# 智能问答系统依赖包
# 数据库连接
mysql-connector-python>=8.0.32

# 数据处理
pandas>=1.5.0
numpy>=1.21.0

# 自然语言处理
jieba>=0.42.1
nltk>=3.7

# 命令行彩色输出
colorama>=0.4.4

# 配置管理
python-dotenv>=0.19.0